Refactor exports in shared/index.ts
This commit explicitly specifies the types and module file extensions for
the exported types and modules in `shared/index.ts`. Also destructures
and re-exports the types and modules.
This was done to address so that the types and modules could be directly
accessed in the backend.
The initial commit had the types and modules exported using the `*`
wildcard which led to issues with the types and modules not being
accessible in the backend.