Add JSON body parsing and add next function error-middleware args
The error middleware was not being called when errors were thrown.
This was because the next function was missing from the args and hence
it wasn't functioning as an error middleware in the first place.

This commit adds `express.json()` middleware for parsing json requests
and adds `next` function in error middleware args.