chore: Remove Toaster component from main entry file and use in routes
The Toaster component needs to be displayed at different position based
on user's screen. Therefore, in order to do some conditional rendering
and to use hooks, it had to be moved to a different component where
it could still be at the top level.
The routes file was the only such place and hence Toaster is being used
there.