feat: Update return type for api client to access api response messages
Initially, the data itself was being returned by the ApiClient fetch.
Because of this the ApiResponse messages remained unused.

This change updates the return type of the ApiClient to ApiResponse type
hence, returning the response itself.

The issue with previous approach was that in order to send toast
notifications to the end-user, we had to evaluate the action type of
each thunk inside the middleware which would've led to long lines of
switch statements.

Because of this change, there is no need to manually send customized
notification by evaluating each thunk. Instead, we simply return the
message from ApiResponse, which is received in a user-friendly format,
hence, massively simplifying the code and keeping things DRY.