feat: Create readonly type discriminator for custom defined error class
The error created from class when thrown, were spread into objects
before being passed into `rejectWithValue` of async thunks. This was
done because actions cannot be a class instance.

Since the error were passed as object, they ended up losing their
prototype chain because of which the `isCustomDefinedError` type guard
failed as `instanceof` checks could not identify the class instance
which was now, actually an object.

To fix this, a readonly property `__type` has been added to both
CustomError and FetchError definitions in order to perfectly identify
them as instances of their classes.