chore: Remove redundant type from Error response generic and instances
The type `Record<string, unknown>` was actually specific to Prisma
generated errors and was ultimately not used anywhere in the codebase.

Also, it is better to not throw Prisma generated errors in frontend but
instead just log them in backend console for debugging. This keeps any
data leakages to a minimum.

The `Record<string, unknown>` has now been removed from everywhere in
the frontend including the `error` property type in custom defined
errors and error redux state or the ApiResponseError generic in
`apiClient`.