feat: Remove component props and get update data from useLocation state
Initially, the data was supposed to be passed as props in case of
trainer update.
However, we are now passing data via react-router navigate as state.

Therefore, the component props have been removed and the data is now
obtained from `useLocation` state.
All type checks pertaining to the same have been made as well.

The location state is cleared on component unmount to prevent data
persistence.