chore: Remove redundant clearError reducer
Initially, the clearError reducer was called to reset error state on
component unmount.
However, this was not required because any new errors can just overwrite
the error state instead, resulting in the same outcome.

An issue with clearError was that on on component rerenders of strict
mode, the errors were cleared, resulting in no ereor messages.