refactor: Remove redundant formSlice redux slice
The form slice was created to handle create/update tasks for trainer
and pokemon. I was under the impression that async thunks could not be
used without a slice but they do work.

No redux state management was needed for handling form related tasks.
Only thunks were needed to make the requests so that middlewares could
do the rest of the task.

Therefore, the module itself has been named to `formThunks.ts` since all
it does is store async thunks used for form related tasks.

The slice was redundant here as there was no state management being
performed and hence has been removed.