refactor: Remove redundant dataSlice redux slice
The data slice was created to handle GET/DELETE tasks for individual
trainer and pokemon. I was under the impression that async thunks could not be
used without a slice but they do work.

No redux state management was needed for handling the data. Only thunks
were needed to make the requests so that middlewares could do the rest
of the task.

Therefore, the module itself has been named to `dataThunks.ts` since all
it does is store async thunks.

The slice was redundant here as there was no state management being
performed and hence has been removed.