feat: Segregate toast handling logic to prevent race condition
When a user swiftly visited routes, the toasts were either not shown or
shown in the wrong order, clearly indicating a race condition.
This was because all the GET type actions, i.e., fetch Trainer/Pokemon
by ID or fetch entire Trainers/Pokemon list, were handled in a sinlge
`if` statement.

To prevent this race condition, we segregated the toast handling logic
to different action types, i.e., list fetch in one statement and
singular (by ID) fetch in another.